Move implementation of convert to cpp file
